+++ extension/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the Tabstronaut extension will be documented in this file
 
 ## [1.2.7]
 
-- Optional automatic grouping of new tabs by file type or top-level folder.
+- New action to sort tabs within a group by folder or file type.
 
 ## [1.2.6]